Tidy up

+++ src/mdf.rs
@@ -198,11 +198,8 @@ impl MDF {
     }
 
     pub fn read_channel(&self, channel: &MdfChannel) -> Signal {
-        self.file.read(
-            channel.data_group as usize,
-            channel.channel_group as usize,
-            channel.channel as usize,
-        )
+        self.file
+            .read(channel.data_group, channel.channel_group, channel.channel)
     }
 }
